+++ keyboards/nullbitsco/snap/keymaps/bongo_reactive/config.h
@@ -18,9 +18,8 @@
 
 /* space savers */
 #define DYNAMIC_KEYMAP_LAYER_COUNT 3
-// #define NO_ACTION_TAPPING // Waiting on QMK #11528
+#define NO_ACTION_TAPPING
 #define NO_ACTION_ONESHOT
-#define IGNORE_MOD_TAP_INTERRUPT
 #define TAPPING_FORCE_HOLD
 
 #define OLED_BRIGHTNESS 128
@@ -29,13 +28,6 @@
 // Selectively undefine to save space
 // VIA support won't fit otherwise
 #ifdef RGBLIGHT_ENABLE
-#undef RGBLIGHT_LED_MAP
-#undef RGBLIGHT_EFFECT_SNAKE
-#undef RGBLIGHT_EFFECT_KNIGHT
-#undef RGBLIGHT_EFFECT_CHRISTMAS
-#undef RGBLIGHT_EFFECT_STATIC_GRADIENT
-#undef RGBLIGHT_EFFECT_RGB_TEST
-#undef RGBLIGHT_EFFECT_ALTERNATING
 #undef RGBLIGHT_EFFECT_TWINKLE
 #endif //RGB LIGHT_ENABLE
