+++ keyboards/nullbitsco/snap/keymaps/oled/config.h
@@ -18,22 +18,10 @@
 
 /* space savers */
 #define DYNAMIC_KEYMAP_LAYER_COUNT 3
-// #define NO_ACTION_TAPPING // Waiting on QMK #11528
+#define NO_ACTION_TAPPING
 #define NO_ACTION_ONESHOT
-#define IGNORE_MOD_TAP_INTERRUPT
 #define TAPPING_FORCE_HOLD
 
 #define OLED_BRIGHTNESS 128
 #define OLED_TIMEOUT 30000
 #define OLED_UPDATE_INTERVAL 250
-
-// Selectively undefine to save space
-// VIA support won't fit otherwise
-#ifdef RGBLIGHT_ENABLE
-#undef RGBLIGHT_EFFECT_SNAKE
-#undef RGBLIGHT_EFFECT_KNIGHT
-#undef RGBLIGHT_EFFECT_CHRISTMAS
-#undef RGBLIGHT_EFFECT_RGB_TEST
-#undef RGBLIGHT_EFFECT_ALTERNATING
-#undef RGBLIGHT_EFFECT_TWINKLE
-#endif //RGB LIGHT_ENABLE
