+++ keyboards/nullbitsco/snap/keymaps/typehud/rules.mk
@@ -3,3 +3,5 @@ OLED_DRIVER = SSD1306
 WPM_ENABLE = yes
 VIA_ENABLE = yes
 ENCODER_MAP_ENABLE = yes
+
+SRC += typehud.c
